ajout services et serviceImp

+++ src/main/java/sn/ahiba/gmembrebacken/entities/Section.java
@@ -26,7 +26,8 @@ public class Section {
         return id;
     }
 
-    public void setId(int id) {
+
+    public void setId(long id) {
         this.id = id;
     }
